Fix spec issue, the code is already doing this.
So there is no point in having these metadata offsets like this as the main purpose of offsets is to be able to jump to a specific thing. This doesn't have anyway to do so and you would need to loop through all of them to find a specific item regardless.